requirements.txt
asgiref==3.6.0
Cerberus==1.3.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-constance==2.9.1
django-filter==23.1
django-picklefield==3.1
django-rest-swagger==2.2.0
django-versatileimagefield==3.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
openapi-codec==1.3.2
Pillow==9.4.0
PyJWT==2.6.0
python-magic==0.4.27
pytz==2023.2
requests==2.28.2
simplejson==3.18.4
sqlparse==0.4.3
toolz==0.12.0
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==1.26.15

django-ckeditor~=6.5.1